refactor: Product 패키지 이동
